requirements.txt
asn1crypto==1.4.0
cffi==1.14.6
cryptography==3.4.8
pkg_resources==0.0.0
pycparser==2.20
pyOpenSSL==20.0.1
six==1.16.0
